Zapytaj o potwierdzenie przed restartem kontenera

+++ edeklaracje.sh
@@ -66,9 +66,17 @@ fi
 # czy przypadkiem nie ma uruchomionego innego dockera z tą samą nazwą?
 docker_state=$(docker inspect -f '{{.State}}' "$CONTAINER_NAME" 2> /dev/null)
 if [[ ! -z "$docker_state" && "$docker_state" != '<no value>' ]]; then
-  echo -ne "Wygląda na to, że kontener $CONTAINER_NAME istnieje; zatrzymuję/niszczę, by móc uruchomić na nowo.\n\n"
-  docker stop "$CONTAINER_NAME"
-  docker rm -v "$CONTAINER_NAME"
+  echo -ne "Wygląda na to, że kontener $CONTAINER_NAME jest uruchomiony. Czy zatrzymać/zniszczyć go, by móc uruchomić na nowo? (T/n) "
+  read restart_container
+  if [[ $restart_container != 'n' && $restart_container != 'N' ]]; then
+    echo -ne '\nZatrzymuję kontener...\n\n'
+    docker stop "$CONTAINER_NAME"
+    docker rm -v "$CONTAINER_NAME"
+    echo -ne '\nKontener zatrzymany.\n\n'
+  else
+    echo -ne 'Anulowano.\n\n'
+    exit 0
+  fi
 fi
 
 # na wszelki wypadek pytamy juzera
